# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Maturity Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type CreateAssessmentInputCompanySize =
  (typeof CreateAssessmentInputCompanySize)[keyof typeof CreateAssessmentInputCompanySize];

export const CreateAssessmentInputCompanySize = {
  startup: "startup",
  small: "small",
  medium: "medium",
  large: "large",
  enterprise: "enterprise",
} as const;

export interface CreateAssessmentInput {
  organizationName: string;
  industry: string;
  companySize: CreateAssessmentInputCompanySize;
  geography: string;
  contactName?: string;
  contactEmail?: string;
}

export type AssessmentStatus =
  (typeof AssessmentStatus)[keyof typeof AssessmentStatus];

export const AssessmentStatus = {
  draft: "draft",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type AssessmentMaturityLevel =
  | (typeof AssessmentMaturityLevel)[keyof typeof AssessmentMaturityLevel]
  | null;

export const AssessmentMaturityLevel = {
  nascent: "nascent",
  emerging: "emerging",
  developing: "developing",
  advanced: "advanced",
  leading: "leading",
} as const;

export interface Assessment {
  id: number;
  organizationName: string;
  industry: string;
  companySize: string;
  geography: string;
  contactName?: string | null;
  contactEmail?: string | null;
  status: AssessmentStatus;
  maturityLevel?: AssessmentMaturityLevel;
  overallScore?: number | null;
  createdAt: string;
  completedAt?: string | null;
}

export interface SurveyResponse {
  id: number;
  assessmentId: number;
  questionId: string;
  category: string;
  numericValue?: number | null;
  textValue?: string | null;
}

export type AssessmentWithResponsesStatus =
  (typeof AssessmentWithResponsesStatus)[keyof typeof AssessmentWithResponsesStatus];

export const AssessmentWithResponsesStatus = {
  draft: "draft",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface AssessmentWithResponses {
  id: number;
  organizationName: string;
  industry: string;
  companySize: string;
  geography: string;
  contactName?: string | null;
  contactEmail?: string | null;
  status: AssessmentWithResponsesStatus;
  maturityLevel?: string | null;
  overallScore?: number | null;
  createdAt: string;
  completedAt?: string | null;
  responses: SurveyResponse[];
}

export interface SurveyAnswer {
  questionId: string;
  category: string;
  numericValue?: number | null;
  textValue?: string | null;
}

export interface SubmitAssessmentInput {
  answers: SurveyAnswer[];
}

export interface CategoryScore {
  category: string;
  categoryLabel: string;
  score: number;
  maxScore: number;
  percentage: number;
}

export type RecommendationTimeframe =
  (typeof RecommendationTimeframe)[keyof typeof RecommendationTimeframe];

export const RecommendationTimeframe = {
  short_term: "short_term",
  mid_term: "mid_term",
  long_term: "long_term",
} as const;

export type RecommendationPriority =
  (typeof RecommendationPriority)[keyof typeof RecommendationPriority];

export const RecommendationPriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type RecommendationEffort =
  (typeof RecommendationEffort)[keyof typeof RecommendationEffort];

export const RecommendationEffort = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type RecommendationImpact =
  (typeof RecommendationImpact)[keyof typeof RecommendationImpact];

export const RecommendationImpact = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Recommendation {
  id: string;
  title: string;
  description: string;
  category: string;
  timeframe: RecommendationTimeframe;
  priority: RecommendationPriority;
  effort: RecommendationEffort;
  impact: RecommendationImpact;
  tools: string[];
}

export interface AssessmentResult {
  assessmentId: number;
  organizationName: string;
  industry: string;
  companySize: string;
  maturityLevel: string;
  maturityLabel: string;
  overallScore: number;
  maxScore: number;
  percentage: number;
  categoryScores: CategoryScore[];
  recommendations: Recommendation[];
  executiveSummary: string;
  strengthAreas: string[];
  gapAreas: string[];
  completedAt: string;
}

export interface DashboardStats {
  totalAssessments: number;
  completedAssessments: number;
  averageScore: number;
  averageMaturityLevel: string;
  assessmentsThisMonth: number;
  topIndustry: string;
}

export interface MaturityDistributionItem {
  level: string;
  label: string;
  count: number;
  percentage: number;
}

export interface IndustryBenchmark {
  industry: string;
  averageScore: number;
  topScore: number;
  assessmentCount: number;
}

export type SurveyQuestionType =
  (typeof SurveyQuestionType)[keyof typeof SurveyQuestionType];

export const SurveyQuestionType = {
  scale: "scale",
  multiple_choice: "multiple_choice",
  text: "text",
} as const;

export type SurveyQuestionOptionsItem = {
  value: number;
  label: string;
};

export interface SurveyQuestion {
  id: string;
  text: string;
  type: SurveyQuestionType;
  options?: SurveyQuestionOptionsItem[] | null;
  weight: number;
  helpText?: string | null;
}

export interface SurveyCategory {
  id: string;
  label: string;
  description: string;
  icon: string;
  questions: SurveyQuestion[];
}