# lib/api-zod/src/generated/types/assessmentResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Maturity Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryScore } from "./categoryScore";
import type { Recommendation } from "./recommendation";

export interface AssessmentResult {
  assessmentId: number;
  organizationName: string;
  industry: string;
  companySize: string;
  maturityLevel: string;
  maturityLabel: string;
  overallScore: number;
  maxScore: number;
  percentage: number;
  categoryScores: CategoryScore[];
  recommendations: Recommendation[];
  executiveSummary: string;
  strengthAreas: string[];
  gapAreas: string[];
  completedAt: Date;
}