# lib/api-zod/src/generated/types/assessmentWithResponses.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Maturity Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentWithResponsesStatus } from "./assessmentWithResponsesStatus";
import type { SurveyResponse } from "./surveyResponse";

export interface AssessmentWithResponses {
  id: number;
  organizationName: string;
  industry: string;
  companySize: string;
  geography: string;
  contactName?: string | null;
  contactEmail?: string | null;
  status: AssessmentWithResponsesStatus;
  maturityLevel?: string | null;
  overallScore?: number | null;
  createdAt: Date;
  completedAt?: Date | null;
  responses: SurveyResponse[];
}