# lib/api-zod/src/generated/types/categoryScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Maturity Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryScore {
  category: string;
  categoryLabel: string;
  score: number;
  maxScore: number;
  percentage: number;
}