# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Maturity Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalAssessments: number;
  completedAssessments: number;
  averageScore: number;
  averageMaturityLevel: string;
  assessmentsThisMonth: number;
  topIndustry: string;
}