# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Maturity Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationEffort } from "./recommendationEffort";
import type { RecommendationImpact } from "./recommendationImpact";
import type { RecommendationPriority } from "./recommendationPriority";
import type { RecommendationTimeframe } from "./recommendationTimeframe";

export interface Recommendation {
  id: string;
  title: string;
  description: string;
  category: string;
  timeframe: RecommendationTimeframe;
  priority: RecommendationPriority;
  effort: RecommendationEffort;
  impact: RecommendationImpact;
  tools: string[];
}