# lib/api-zod/src/generated/types/submitAssessmentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Maturity Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SurveyAnswer } from "./surveyAnswer";

export interface SubmitAssessmentInput {
  answers: SurveyAnswer[];
}