# lib/api-zod/src/generated/types/surveyQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Maturity Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SurveyQuestionOptionsItem } from "./surveyQuestionOptionsItem";
import type { SurveyQuestionType } from "./surveyQuestionType";

export interface SurveyQuestion {
  id: string;
  text: string;
  type: SurveyQuestionType;
  options?: SurveyQuestionOptionsItem[] | null;
  weight: number;
  helpText?: string | null;
}