# lib/api-zod/src/generated/types/surveyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Maturity Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SurveyResponse {
  id: number;
  assessmentId: number;
  questionId: string;
  category: string;
  numericValue?: number | null;
  textValue?: string | null;
}